io/command: Fix test for invalid GPid pid
Meanwhile GPid is a HANDLE (a void pointer) for Windows.
Therefore code like pid > 0 is no longer valid for Windows
and causes a compiler warning.

GPid is still an integer value for Linux.

As the initial value (invalid pid) also changed from -1 to 0,
testing for 0 instead of testing for pid > 0 can now be used
for Linux and for Windows to detect an invalid pid.

Fixes: a95570e3e4d61a7c ("io/command: use glib GSpawn [...]")
Signed-off-by: Stefan Weil <sw@weilnetz.de>